fix: unifique styles
